requirements.txt
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
pdoc3==0.9.2
python-decouple==3.8
black==22.3.0
autoflake==1.4
twine==4.0.0
-e .
